refactor: split built-in callouts out of constants
