+++ src/constants.ts
@@ -1,20 +1,26 @@
 /**
- * constants.ts — Global constants and default callout definitions.
+ * constants.ts — Global constants and plugin defaults.
  *
  * Exports compile-time constants (MAX_TAG_LENGTH, HEAVY_VAULT_FILE_THRESHOLD,
- * DEFAULT_SETTINGS) and the DEFAULT_CALLOUTS array that seeds the registry with
- * Obsidian's built-in callout types on first load.
+ * PLUGIN_ICON_ID) and DEFAULT_SETTINGS. The 13 built-in callout definitions
+ * they used to sit beside now live in `defaultCallouts.ts`.
  * Imported by CalloutRegistry (for defaults), main.ts (for threshold checks),
  * and several utility/settings modules.
  */
 import type {
-	CalloutDefinition,
 	CalloutIcon,
 	CalloutRenderRole,
 	ContextMenuItemConfig,
 	GlobalStyleSettings,
 	PluginSettings,
 } from "./types";
+
+/**
+ * The 13 built-ins now live in their own module — this file was carrying the
+ * plugin's constants and a 159-line data table at once. Re-exported here so the
+ * registry, the discovered-row builder and the suites keep one import path.
+ */
+export { DEFAULT_CALLOUTS } from "./defaultCallouts";
 
 /**
  * Generous safety cap on callout-ID length, applied only when validating
@@ -25,6 +31,16 @@ export const MAX_TAG_LENGTH: number = 200;
 
 /** Maximum number of IDs (aliases) allowed per callout */
 export const MAX_TAGS_COUNT: number = 4;
+
+/**
+ * The Lucide id that stands for Callout Studio itself.
+ *
+ * Named once because two surfaces wear it and they must not drift: the welcome
+ * splash's hero mark and the left-ribbon button. The plugin registers no icon
+ * of its own with `addIcon()` — this is a stock Obsidian glyph, so it needs no
+ * asset, costs no bytes and follows the theme.
+ */
+export const PLUGIN_ICON_ID = "paintbrush";
 
 /**
  * Placeholder callout ID the editor's live preview renders under while the
@@ -105,178 +121,6 @@ export const OBSIDIAN_CALLOUT_VAR: Readonly<Record<string, string>> =
 		example: "--callout-example",
 		quote: "--callout-quote",
 	});
-
-/**
- * The 13 built-ins, coloured to match Obsidian.
- *
- * The hexes are the resolved values of {@link OBSIDIAN_CALLOUT_VAR} in core's
- * default theme, read out of the shipped `app.css` (`--color-blue` and friends
- * under `.theme-light` / `.theme-dark`). They are a *seed*, not the rendering
- * path: an untouched built-in never has its `--callout-color` overridden at all,
- * so what actually paints is the theme's. These values are what the editor
- * shows in its swatches, what a customized copy starts from, and what
- * `resetBuiltIn` restores.
- *
- * This is the first time a built-in's light and dark accents differ — Obsidian's
- * own palette shifts between themes (its `example` purple moves by ΔE00 15.4),
- * and carrying one value for both was a large part of why the plugin's colours
- * never quite matched. Backgrounds are still deliberately absent: a definition
- * with no `bgColorLight`/`bgColorDark` emits no background rule, which leaves
- * core's translucent fill in place and is what lets nested callouts step.
- */
-export const DEFAULT_CALLOUTS: CalloutDefinition[] = [
-	{
-		id: "note",
-		displayName: "Note",
-		icon: { type: "lucide", value: "pencil" },
-		colorLight: "#086ddd",
-		colorDark: "#027aff",
-		foldable: false,
-		defaultFolded: false,
-		builtIn: true,
-		source: "builtin",
-	},
-	{
-		id: "abstract",
-		displayName: "Abstract",
-		icon: { type: "lucide", value: "clipboard-list" },
-		colorLight: "#00bfbc",
-		colorDark: "#53dfdd",
-		foldable: false,
-		defaultFolded: false,
-		builtIn: true,
-		source: "builtin",
-		aliases: ["summary", "tldr"],
-	},
-	{
-		id: "info",
-		displayName: "Info",
-		icon: { type: "lucide", value: "info" },
-		colorLight: "#086ddd",
-		colorDark: "#027aff",
-		foldable: false,
-		defaultFolded: false,
-		builtIn: true,
-		source: "builtin",
-	},
-	{
-		id: "todo",
-		displayName: "Todo",
-		icon: { type: "lucide", value: "check-circle-2" },
-		colorLight: "#086ddd",
-		colorDark: "#027aff",
-		foldable: false,
-		defaultFolded: false,
-		builtIn: true,
-		source: "builtin",
-	},
-	{
-		id: "tip",
-		displayName: "Tip",
-		icon: { type: "lucide", value: "flame" },
-		colorLight: "#00bfbc",
-		colorDark: "#53dfdd",
-		foldable: false,
-		defaultFolded: false,
-		builtIn: true,
-		source: "builtin",
-		aliases: ["hint", "important"],
-	},
-	{
-		id: "success",
-		displayName: "Success",
-		icon: { type: "lucide", value: "check" },
-		colorLight: "#08b94e",
-		colorDark: "#44cf6e",
-		foldable: false,
-		defaultFolded: false,
-		builtIn: true,
-		source: "builtin",
-		aliases: ["check", "done"],
-	},
-	{
-		id: "question",
-		displayName: "Question",
-		icon: { type: "lucide", value: "help-circle" },
-		colorLight: "#ec7500",
-		colorDark: "#e9973f",
-		foldable: false,
-		defaultFolded: false,
-		builtIn: true,
-		source: "builtin",
-		aliases: ["help", "faq"],
-	},
-	{
-		id: "warning",
-		displayName: "Warning",
-		icon: { type: "lucide", value: "alert-triangle" },
-		colorLight: "#ec7500",
-		colorDark: "#e9973f",
-		foldable: false,
-		defaultFolded: false,
-		builtIn: true,
-		source: "builtin",
-		aliases: ["caution", "attention"],
-	},
-	{
-		id: "failure",
-		displayName: "Failure",
-		icon: { type: "lucide", value: "x" },
-		colorLight: "#e93147",
-		colorDark: "#fb464c",
-		foldable: false,
-		defaultFolded: false,
-		builtIn: true,
-		source: "builtin",
-		aliases: ["fail", "missing"],
-	},
-	{
-		id: "danger",
-		displayName: "Danger",
-		icon: { type: "lucide", value: "zap" },
-		colorLight: "#e93147",
-		colorDark: "#fb464c",
-		foldable: false,
-		defaultFolded: false,
-		builtIn: true,
-		source: "builtin",
-		aliases: ["error"],
-	},
-	{
-		id: "bug",
-		displayName: "Bug",
-		icon: { type: "lucide", value: "bug" },
-		colorLight: "#e93147",
-		colorDark: "#fb464c",
-		foldable: false,
-		defaultFolded: false,
-		builtIn: true,
-		source: "builtin",
-	},
-	{
-		id: "example",
-		displayName: "Example",
-		icon: { type: "lucide", value: "list" },
-		colorLight: "#7852ee",
-		colorDark: "#a882ff",
-		foldable: false,
-		defaultFolded: false,
-		builtIn: true,
-		source: "builtin",
-	},
-	{
-		id: "quote",
-		displayName: "Quote",
-		icon: { type: "lucide", value: "quote" },
-		colorLight: "#9e9e9e",
-		colorDark: "#9e9e9e",
-		foldable: false,
-		defaultFolded: false,
-		builtIn: true,
-		source: "builtin",
-		aliases: ["cite"],
-	},
-];
 
 /**
  * Default right-click menu layout per render role. Array order = menu order.
