+++ tests/repoSourceRules.test.ts
@@ -633,7 +633,6 @@ describe("no new oversized files", () => {
 		"src/icons/packs/materialFont.ts": 398,
 		"src/settings/sections/CalloutRowActions.ts": 396,
 		"src/outline/OutlineDecorator.ts": 382,
-		"src/constants.ts": 379,
 		"src/settings/EmbeddableMarkdownEditor.ts": 377,
 		"src/icons/svg.ts": 367,
 		"src/settings/iconpicker/ImagePanel.ts": 354,
